feat: initial move of liveness->dysfunctional detection crate in root
