+++ dys_function/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+authors = ["MaidSafe Developers <dev@maidsafe.net>"]
+description = "The Safe Network Node Dysfunctional Detection."
+documentation = "https://docs.rs/sn_node"
+edition = "2021"
+homepage = "https://maidsafe.net"
+license = "GPL-3.0"
+name = "dys_function"
+readme = "README.md"
+repository = "https://github.com/maidsafe/safe_network"
+version = "0.1.0"
+
+
+[features]
+default = []
+
+
+[dependencies]
+xor_name = "4.0.1"
+dashmap = {version = "5.1.0", features = [ "serde" ]}
+itertools = "0.10.0"
+tracing = "~0.1.26"
+rand = "~0.8"
+
+[dependencies.tokio]
+version = "1.17.0"
+features = ["sync"]
+
+[dev-dependencies]
+eyre = "~0.6.5"
+
+[dev-dependencies.tokio]
+version = "1.17.0"
+features = ["sync", "macros", "rt-multi-thread"]
