+++ sn/Cargo.toml
@@ -90,6 +90,7 @@ url = "2.2.0"
 walkdir = "2"
 xor_name = "4.0.1"
 file-rotate = "~0.6.0"
+dys_function = { path = "../dys_function", version = "0.1.0" }
 
 [dependencies.backoff]
 version = "0.4.0"
